Add health checks to initial game setup
When the game initially is setup the api-load action performs a health check.

In effect snakes that respond are flagged are marked as healthy, and snakes
with no, or bad, response are flagged as unhealthy.

At the moment this means that unhealthy snakes will not cause the game to crash.
In the future I'll likely make the game refuse to start if there are unhealthy
snakes.