Refactor GameServer
Reorganized function definitions in GameServer with the intention of moving
state of {status, GameServer.State} into GameServer.State.state.

To this end all server_callbacks are grouped under their request atom and the
status matching is done in a case statement in the function definition.

This does also make reading this file significantly easier.