Optimize dom replacement for board viewer
This one line cuts down the "loading" slice from the chrome timeline from 25% to
3%. jQuery.html performs a very expensive html validation before inserting to
into the dom.

Because we *hope* that we're sending valid html before hand we don't need to do
this.

This leaves page redraw and paint as the next largest bottleneck, which cannot
be overcome without using something other than svg.