Fix issue where game server would fail to connect
the find method here would attempt to start the server if it wasn't running.

There was a data race where another process could also attempt to start it.

In the even that another process started it this just returns the pid, which
was what it was supposed to do in the first place.