Fix mnesia host config for prod releases
There's a lot of mismatched between where mnesia_host is expected to be
defined.

EctoMnesia bakes it into the build on compiliation, which is usually
nonode@nohost, but it will also pull it from the repo's conifg.

The easiest option is to just hardcode the host name to that of the
host.