Update dockerfile
The current alpine build fails on elm-make, because elm-make doesn't work in a musl based libc environment which alpine is. This starts from a glibc based alpine image, then builds up first erlang, then elixir, then the battlesnake server. The final image can just be alpine based, because we don't need elm-make in that image since it was compiled down to static assets.